requirements.txt
# Requirements for Preprint_Repository
# 
# Installation:
#   pip install -r requirements.txt

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Deep learning (CPU version by default, see instructions above for GPU)
torch>=1.9.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Reinforcement learning environments
gymnasium>=0.28.0

# Serialization
dill>=0.3.4

# Performance optimization
numba>=0.54.0

# Interactive computing (optional)
ipython>=7.0.0
